Py qt 6

+++ server_storage.py
@@ -10,9 +10,11 @@ from datetime import datetime
 
 class ServerStorage:
     class AllUser:
-        def __init__(self, name, last_login_date):
+        def __init__(self, name, passwrd, last_login_date):
             self.name = name
+            self.passwrd = passwrd
             self.last_login_date = last_login_date
+
             self.id = None
 
         def __repr__(self):
@@ -48,6 +50,15 @@ class ServerStorage:
         def __repr__(self):
             return f'{self.user_1_id} - {self.user_2_id}'
 
+    class UserKey:
+        def __init__(self, user_id, public_key):
+            self.user_id = user_id
+            self.public_key = public_key
+            self.id = None
+
+        def __repr__(self):
+            return f'User - {self.user_id}'
+
     def __init__(self):
         # engine = create_engine('sqlite:///server_db.db3?check_same_thread=False', echo=False)
         config = configparser.ConfigParser()
@@ -64,6 +75,7 @@ class ServerStorage:
         clients_table = Table('all_users', self.meta_data,
                               Column('id', Integer, primary_key=True),
                               Column('name', String),
+                              Column('passwrd', String),
                               Column('last_login_date', DateTime))
 
         active_users = Table('active_users', self.meta_data,
@@ -85,23 +97,30 @@ class ServerStorage:
                          Column('user_2_id', Integer, ForeignKey('all_users.id'))
                          )
 
+        keys = Table('keys', self.meta_data,
+                         Column('id', Integer, primary_key=True),
+                         Column('user_id', Integer, ForeignKey('all_users.id')),
+                         Column('public_key', String)
+                         )
+
         self.meta_data.create_all(engine)
         mapper(self.AllUser, clients_table)
         mapper(self.ActiveUser, active_users)
         mapper(self.LoginHistory, login_history)
         mapper(self.Contact, contacts)
+        mapper(self.UserKey, keys)
         self.clear_active_users()
 
-    def create_user(self, name):
-        """Добавляем нового пользователя, если он есть меняем дату последнего входа """
+    def create_user(self, name,passwrd):
+        """Добавляем нового пользователя, если он есть возвращаем False """
         user = self.session.query(self.AllUser).filter_by(name=name).first()
         if user:
-            user.last_login_date = datetime.now()
+            return False
         else:
-            user = self.AllUser(name, datetime.now())
-        self.session.add(user)
-        self.session.commit()
-        return user
+            user = self.AllUser(name,passwrd, datetime.now())
+            self.session.add(user)
+            self.session.commit()
+            return True
 
     def delete_user(self, name):
         """Удаляем пользователя"""
@@ -120,22 +139,28 @@ class ServerStorage:
         return user
 
     def user_login(self, name, ip, port):
-        """ Создает нового пользователя, или меняет дату входа у старого,
+        """
         добавлет в активные и заносит в журнал истории логинов"""
-        new_user = self.create_user(name)
-        self.add_new_active_user(new_user.id, ip, port)
-        history = self.LoginHistory(new_user.id, ip, port)
+        user = self.session.query(self.AllUser).filter_by(name=name).first()
+        self.add_new_active_user(user.id, ip, port)
+        history = self.LoginHistory(user.id, ip, port)
         self.session.add(history)
         self.session.commit()
 
+    def get_user_pass(self,name):
+        user = self.session.query(self.AllUser).filter_by(name=name).first()
+        if user:
+            return user.passwrd
+        else:
+            return False
+
     def delete_active_user(self, user_name):
         ''' Удаляет пользователя из активных '''
-        # user = self.session.query(self.AllUser).filter_by(name=user_name).first()
-        # user = self.session.query(self.ActiveUser).filter_by(user_id=user.id).first()
-        user = \
-            self.session.query(self.AllUser, self.ActiveUser).filter_by(name=user_name).join(self.ActiveUser).first()[1]
-        self.session.delete(user)
-        self.session.commit()
+        if user_name:
+            user = \
+                self.session.query(self.AllUser, self.ActiveUser).filter_by(name=user_name).join(self.ActiveUser).first()[1]
+            self.session.delete(user)
+            self.session.commit()
 
     def clear_active_users(self):
         """ Очищает таблицу активных юзеров"""
@@ -201,35 +226,46 @@ class ServerStorage:
         result = [contact[1].name for contact in user_contacts]
         return result
 
+    def set_key(self,user_name,public_key):
+        user = self.session.query(self.AllUser).filter_by(name=user_name).first()
+        key = self.session.query(self.UserKey).filter_by(user_id = user.id).first()
+        if key:
+            if key.public_key != public_key:
+                key.public_key = public_key
+        else:
+            key = self.UserKey(user.id,public_key)
+        self.session.add(key)
+        self.session.commit()
 
 if __name__ == '__main__':
     server = ServerStorage()
-    new_user = server.create_user('Vovas')
-    server.add_new_active_user(new_user.id, '127.0.0.1', '7777')
-    new_user_2 = server.create_user('Dima')
-    server.add_new_active_user(new_user_2.id, '127.0.0.1', '7777')
-
-    server.user_login('Vanya', '127.0.0.1', '8888')
-    full_data = server.session.query(server.AllUser.name, server.ActiveUser.ip, server.AllUser.last_login_date).join(
-        server.AllUser).all()
-    # print(full_data)
-    server.user_login('Petya', '127.0.0.1', '8888')
-    print('-------------')
-    print(server.get_active_users(name='Vovas'))
+    server.set_key('User-1','123')
+    # new_user = server.create_user('Vovas','Paasword')
+    # server.add_new_active_user(new_user.id, '127.0.0.1', '7777')
+    # new_user_2 = server.create_user('Dima','Passwrd2')
+    # server.add_new_active_user(new_user_2.id, '127.0.0.1', '7777')
+    #
+    # # server.user_login('Vanya', '127.0.0.1', '8888')
+    # full_data = server.session.query(server.AllUser.name, server.ActiveUser.ip, server.AllUser.last_login_date).join(
+    #     server.AllUser).all()
+    # # print(full_data)
+    # # server.user_login('Petya', '127.0.0.1', '8888')
+    # print('-------------')
+    # print(server.get_active_users(name='Vovas'))
     # server.delete_active_user('Vovas')
-
-    query = server.session.query(server.AllUser, server.ActiveUser).outerjoin(server.ActiveUser)
-    query = server.session.query(server.AllUser, server.ActiveUser).outerjoin(server.ActiveUser)
+    #
+    # query = server.session.query(server.AllUser, server.ActiveUser).outerjoin(server.ActiveUser)
+    # query = server.session.query(server.AllUser, server.ActiveUser).outerjoin(server.ActiveUser)
     # print(query)
-    records = query
+    # records = query
     # for user, active_user in records:
     #     # print(user.name,active_user.ip,active_user.port,user.last_login_date)
     #     print(user.name,
     #           f'Подключился {datetime.strftime(user.last_login_date, "%d.%m.%Y %H:%M")}' if active_user else 'Не активен')
     #     # print(user,active_user)
 
-    resp = server.add_new_contact('User-2', 'User-5')
-    print(resp)
-    resp = server.delete_new_contact('User-2', 'User-5')
-    print(resp)
-    contacts = server.get_contacts('User-17')
+    # resp = server.add_new_contact('User-2', 'User-5')
+    # print(resp)
+    # resp = server.delete_new_contact('User-2', 'User-5')
+    # print(resp)
+    # contacts = server.get_contacts('User-17')
